update test imports and helper, remove trial_days from DTO, start offboarding the consts
