+++ shared/plan/constants.py
@@ -101,7 +101,6 @@ class PlanData:
             "benefits": self.benefits,
             "tier_name": self.tier_name,
             "monthly_uploads_limit": self.monthly_uploads_limit,
-            "trial_days": self.trial_days or TrialDaysAmount.CODECOV_SENTRY.value,
             "is_free_plan": self.tier_name == TierName.BASIC.value,
             "is_pro_plan": self.tier_name == TierName.PRO.value,
             "is_team_plan": self.tier_name == TierName.TEAM.value,
@@ -120,7 +119,6 @@ def convert_to_DTO(plan) -> dict:
         "benefits": plan.benefits,
         "tier_name": plan.tier.tier_name,
         "monthly_uploads_limit": plan.monthly_uploads_limit,
-        "trial_days": TrialDaysAmount.CODECOV_SENTRY.value,
         "is_free_plan": not plan.paid_plan,
         "is_pro_plan": plan.tier.tier_name == TierName.PRO.value,
         "is_team_plan": plan.tier.tier_name == TierName.TEAM.value,
