+++ tests/unit/upload/test_utils.py
@@ -18,7 +18,7 @@ from shared.upload.utils import (
     insert_coverage_measurement,
     query_monthly_coverage_measurements,
 )
-from tests.unit.plan.test_plan import mock_all_plans_and_tiers
+from tests.helper import mock_all_plans_and_tiers
 
 
 class CoverageMeasurement(TestCase):
